improve display while popup loads
