+++ public/popup.html
@@ -8,7 +8,7 @@
     <link id="sponsorBlockStyleSheet" href="popup.css" rel="stylesheet">
   </head>
 
-  <body id="sponsorBlockPopupBody">
+  <body id="sponsorBlockPopupBody" style="visibility: hidden">
     <div id="sponsorblockPopup" class="sponsorBlockPageBody sb-preload">
 
       <button id="sbCloseButton" title="__MSG_closePopup__" class="sbCloseButton">
