+++ src/content.ts
@@ -1694,7 +1694,7 @@ function openInfoMenu() {
     const frame = document.createElement("iframe");
     frame.width = "374";
     frame.height = "500";
-    frame.onload = () => frame.contentWindow.postMessage("", "*");
+    frame.addEventListener("load", () => frame.contentWindow.postMessage("", "*"));
     frame.src = chrome.extension.getURL("popup.html");
     popup.appendChild(frame);
