+++ src/utils/pageUtils.ts
@@ -62,3 +62,28 @@ export function getHashParams(): Record<string, unknown> {
 
     return {};
 }
+
+export function localizeHtmlPage(): void {
+    //Localize by replacing __MSG_***__ meta tags
+    const localizedMessage = getLocalizedMessage(document.title);
+    if (localizedMessage) document.title = localizedMessage;
+    const objects = document.getElementsByClassName("sponsorBlockPageBody")[0].children;
+    for (let j = 0; j < objects.length; j++) {
+        const obj = objects[j];
+        const localizedMessage = getLocalizedMessage(obj.innerHTML.toString());
+        if (localizedMessage) obj.innerHTML = localizedMessage;
+    }
+}
+
+export function getLocalizedMessage(text: string): string | false {
+    const valNewH = text.replace(/__MSG_(\w+)__/g, function(match, v1) {
+        return v1 ? chrome.i18n.getMessage(v1).replace(/</g, "&#60;")
+            .replace(/"/g, "&quot;").replace(/\n/g, "<br/>") : "";
+    });
+
+    if (valNewH != text) {
+        return valNewH;
+    } else {
+        return false;
+    }
+}
